+++ FE/src/components/SettingContent/CalendarSetting/index.tsx
@@ -2,7 +2,9 @@ import React, { useState, useEffect } from 'react';
 
 import { useAccountBookData } from '../../../store/AccountBook/accountBookInfoHook';
 
-import CalendarHeader from '../../Calendar/CalendarHeader';
+import DayRadioButton from './DayRadioButton';
+import WeekHeader from './WeekHeader';
+import ActionButton from '../../Common/ActionButton';
 
 import { updateStartDay } from '../../../api/calendar-startday';
 import './calendarSetting.scss';
@@ -16,7 +18,6 @@ export default function CalendarSetting({
   const start = useAccountBookData(store => store.accountBook.startday);
 
   const [startDay, setStartDay] = useState('');
-  const weeks = ['SUN', 'MON', 'TUE', 'WED', 'THU', 'FRI', 'SAT'];
 
   const onClickSaveBtn = () => {
     setSaveModal(() => true);
@@ -34,32 +35,11 @@ export default function CalendarSetting({
     <div className="calendar__setting__container">
       <div className="calendar__setting__startday">
         <div className="calendar__setting__title">Start Day</div>
-        <div className="select__startday">
-          {weeks.map(day => {
-            return (
-              <label key={day}>
-                {day}
-                <input
-                  type="radio"
-                  name="startday"
-                  value={day}
-                  checked={startDay === day}
-                  onChange={e => setStartDay(e.target.value)}
-                />
-              </label>
-            );
-          })}
+        <DayRadioButton startDay={startDay} setStartDay={setStartDay} />
+        <WeekHeader startDay={startDay} />
+        <div className="calendar__save__btn">
+          <ActionButton type="general" content="Save" action={onClickSaveBtn} />
         </div>
-        <table className="start__day__title">
-          <thead>
-            <tr>
-              <CalendarHeader startDay={startDay} />
-            </tr>
-          </thead>
-        </table>
-        <button className="calendar__setting__save" onClick={onClickSaveBtn}>
-          Save
-        </button>
       </div>
     </div>
   );
